+++ api/extensions/documentation/documentation/1.0.0/full_documentation.json
@@ -14,7 +14,7 @@
       "name": "Apache 2.0",
       "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
     },
-    "x-generation-date": "07/28/2021 3:24:54 PM"
+    "x-generation-date": "08/04/2021 5:07:39 PM"
   },
   "x-strapi-config": {
     "path": "/documentation",
